requirements.txt
asgiref==3.4.1
astroid==2.8.2
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.1
djangorestframework==3.12.4
gunicorn==20.1.0
install==1.3.5
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
